% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{anchorData1}
\alias{anchorData1}
\title{Anchor data 1}
\format{
A named list containing the following elements:
\describe{
\item{internalTrain}{A data frame of internal training set}
\item{internalTest}{A data frame of internal test set}
\item{externalTest}{A data frame of external test set}
}
Each dataset contains the following:
\describe{
\item{X1, ... , X10}{features}
\item{Y}{outcome}
}
}
\usage{
anchorData1
}
\description{
Anchor data 1
}
\keyword{datasets}
