% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{anchorLR1}
\alias{anchorLR1}
\title{Anchor logistic regression model 1}
\format{
A logistic regression model trained using anchorData1::internalTrain
}
\usage{
anchorLR1
}
\description{
Anchor logistic regression model 1
}
\keyword{datasets}
