% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dualSGDWeightOptimizer.R
\name{balanceClassesByUpsampling}
\alias{balanceClassesByUpsampling}
\title{Generate information for class balancing during training and weight coputations}
\usage{
balanceClassesByUpsampling(y)
}
\arguments{
\item{y}{}
}
\value{
a named list with \code{sampleIdx} and \code{nDup}
}
\description{
Generate information for class balancing during training and weight coputations
}
