% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{binaryAnchorLR1}
\alias{binaryAnchorLR1}
\title{Anchor logistic regression model 1 with binary covariates}
\format{
A logistic regression model trained using binaryAnchorData1::internalTrain
}
\usage{
binaryAnchorLR1
}
\description{
Anchor logistic regression model 1 with binary covariates
}
\keyword{datasets}
