% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preDiagnostics.R
\name{checkVariableNamesOverlap}
\alias{checkVariableNamesOverlap}
\title{Check overlap of variables names}
\usage{
checkVariableNamesOverlap(z, mu)
}
\arguments{
\item{z}{a data-frame}

\item{mu}{named vector}
}
\description{
Check overlap of variables names
}
