% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SMDFromTable1LikeTransform.R
\name{computeMaxSMD}
\alias{computeMaxSMD}
\title{Compute maximal standardized mean difference}
\usage{
computeMaxSMD(mu, Z, w = NULL)
}
\arguments{
\item{mu}{statistics}

\item{Z}{data}

\item{w}{(optional) weights}
}
\value{
Returns the maximal SMD
}
\description{
Compute maximal standardized mean difference (SMD) between two sets where the first set is represented
by a vector of statistics and the second set by a possibly weighted sample of transformed data.
}
\details{
the data is assumed to have been generated by table 1 like transformation function
}
