% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EvaluationFromExternalStats.R
\name{createExternalEstimatorSettings}
\alias{createExternalEstimatorSettings}
\title{Create settings for \code{estimateExternalPerformanceFromStatistics}}
\usage{
createExternalEstimatorSettings(
  divergence = "entropy",
  lambda = 1e-06,
  minSd = 1e-04,
  minW = 1e-06,
  distance = "l2",
  optimizationMethod = "primal",
  solver = "ECOS",
  nMaxReweight = 10000,
  nRepetitions = 1,
  stratifiedSampling = T,
  maxProp = 500,
  maxDiff = 0.01,
  maxWSMD = 0.2,
  outputDir = getwd(),
  maxCores = 1
)
}
\arguments{
\item{divergence}{'entropy' or 'chi2'.
'entropy' directs the algorithm to minimize the negative entropy, \eqn{\sum_i w_i \log w_i}.
'chi2' is \eqn{\sum_i(w_i-\frac{1}{n})^2}}

\item{lambda}{lambda - regularization parameter}

\item{minSd}{minimum variance for a columns to be included}

\item{minW}{minimum weight}

\item{distance}{distance between mean vectors, l1 or l2}

\item{optimizationMethod}{primal or dual. Currently dual works only with entropy divergence}

\item{solver}{solver ("ECOS" "ECOS_BB" "SCS" "OSQP")}

\item{nMaxReweight}{maximum number of samples for re-weighting. If this number is smaller than the number of
observations, then the estimator acts on sub-samples of the data-set.}

\item{nRepetitions}{number of repetitions with sub-samples.}

\item{stratifiedSampling}{attempt to sample a balanced number of samples for both outcome groups.}

\item{maxProp}{maximum proportion between external and internal means.}

\item{maxDiff}{maximum difference between  external and internal fequencies in case of unary variables or
binary variables with high proportion.}

\item{maxWSMD}{maximum allowed weighted standardized mean difference}

\item{outputDir}{output directory for logging}

\item{maxCores}{maximum number of cores for parallel processing of multiple sub-samples.}
}
\value{
An object of class \code{externalEstimatorSettings}
}
\description{
evaluation from external stats
}
