% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ReweightDB.R
\name{createReweightSettings}
\alias{createReweightSettings}
\title{create reweight settings}
\usage{
createReweightSettings(
  divergence = "entropy",
  lambda = 1e-06,
  minSd = 1e-04,
  minW = 1e-06,
  distance = "l2",
  optimizationMethod = "primal",
  solver = "ECOS"
)
}
\arguments{
\item{divergence}{distributional divergence}

\item{lambda}{regularization}

\item{minSd}{minimum standard deviation of features}

\item{minW}{minimum weight}

\item{distance}{distance between mean vectors}

\item{optimizationMethod}{primal or dual}

\item{solver}{solver}
}
\value{
An object of class \code{reweightSettings}
}
\description{
create reweight settings
}
