% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cvxWeightOptimizer.R
\name{cvxWeightOptimizer}
\alias{cvxWeightOptimizer}
\title{a reweighting object.}
\usage{
cvxWeightOptimizer(
  lambda = 0.1,
  minSd = 1e-04,
  divergence = "entropy",
  optimizationMethod = "dual"
)
}
\arguments{
\item{lambda}{lambda - regularization parameter}

\item{minSd}{minimum variance for a columns to be included}

\item{divergence}{'entropy' or 'chi2'.
'entropy' directs the algorithm to minimize the negative entropy,
\eqn{-\sum_i w_i \log w_i}.
'chi2' is \eqn{\sum_i{w_i-\frac{1}{n}}**2}}

\item{optimizationMethod}{primal or dual. Currently dual works only with entropy divergence}
}
\value{
An object of class cvxWeightOptimizer
}
\description{
a wrapper for CVX base weight optimizer
}
