% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LearningWithExternalStats.R
\docType{data}
\name{dIntTrain}
\alias{dIntTrain}
\title{Internal train}
\format{
A data frame containing the following elements:
\describe{
\item{X1, ... , X10}{features}
\item{Y}{outcome}
}
}
\usage{
load(file=system.file('data/internalTrain.RData', package = "LearningWithExternalStats"))
TODO change this
}
\description{
Internal train
}
\keyword{datasets}
