% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EvaluationFromExternalStats.R
\name{estimateExternalPerformanceFromStatistics}
\alias{estimateExternalPerformanceFromStatistics}
\title{evaluation from external stats
Estimate external performance from statistics}
\usage{
estimateExternalPerformanceFromStatistics(
  internalData,
  externalStats,
  reweightSettings,
  maxProp = 500,
  nboot = 10
)
}
\arguments{
\item{internalData}{a list that includes internal data and predictions with the following fields:
z: a data frame of transformed feature-outcome pairs
y: a vector of outcomes
p: a vector of predicted outcome probabilities}

\item{externalStats}{a vector of means of transformed feature-outcome pairs

Reweighing algorithm parameters:}

\item{reweightSettings}{an object of class \code{reweightSettings}}

\item{maxProp}{maximum proportion between external and internal means}

\item{nboot}{number of bootstrap repetitions for confidence interval assessment}
}
\value{
a named list with the following fields:
summary: a name list with pre-weighting diagnostics, post-weighting diagnostics, estimated performance measures
and bootstrap based performance measures.
preDiagnosis: more detailed pre-reweighing diagnosis information.
bootstrap: a data frame of results of single bootstrap repetitions
}
\description{
Estimate external performance using external statistics and an internal dataset.
This function reweights the internal z matrix with the objective to make the weighted means as close as
possible to the external means represented by mu. Performance measures are estimated using the resulting
weights, the vector of actual outcomes y, and the predicted outcome probabilities p.
}
