% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EvaluationFromExternalStats.R
\name{estimateExternalPerformanceFromStats}
\alias{estimateExternalPerformanceFromStats}
\title{evaluation from external stats
Estimate external performance from statistics}
\usage{
estimateExternalPerformanceFromStats(
  internalData,
  externalStats,
  divergence = "entropy",
  lambda = 1e-06,
  minSd = 1e-04,
  minW = 1e-06,
  distance = "l2",
  optimizationMethod = "primal",
  maxProp = 500,
  nboot = 10
)
}
\arguments{
\item{internalData}{a list that includes internal data and predictions with the following fields:
z: a data frame of transformed feature-outcome pairs
y: a vector of outcomes
p: a vector of predicted outcome probabilities}

\item{externalStats}{a vector of means of transformed feature-outcome pairs

Reweighing algorithm parameters:}

\item{divergence}{'entropy' or 'chi2'.
'entropy' directs the algorithm to minimize the negative entropy, \eqn{-\sum_i w_i \log w_i}.
'chi2' is \eqn{\sum_i{w_i-\frac{1}{n}}**2}}

\item{lambda}{lambda - regularization parameter}

\item{minSd}{minimum variance for a columns to be included}

\item{optimizationMethod}{primal or dual. Currently dual works only with entropy divergence}

\item{maxProp}{maximum proportion between external and internal means}

\item{nboot}{number of bootstrap repetitions for confidence interval assessment}
}
\value{
a named list with the following fields:
summary: a name list with pre-weighting diagnostics, post-weighting diagnostics, estimated performance measures
and bootstrap based performance measures.
preDiagnosis: more detailed pre-reweighing diagnosis information.
bootstrap: a data frame of results of single bootstrap repetitions

This function reweights the internal z matrix with the objective to make the weighted means as close as
possible to the external means represented by mu. Performance measures are estimated using the resulting
weights, the vector of actual outcomes y, and the predicted outcome probabilities p.

check the type of internal data, if in plp format, convert to
}
\description{
Estimate external performance using external statistics and an internal dataset
}
