% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EvaluationFromExternalStats.R
\name{estimateInternalPerformance}
\alias{estimateInternalPerformance}
\title{Estimate internal performance}
\usage{
estimateInternalPerformance(y, p, nboot)
}
\arguments{
\item{y}{a binary outcome vector}

\item{p}{probabilities vector}

\item{nboot}{number of bootstrap repetitions}
}
\description{
estimate internal performance of a model with bootstrap based confidence interval
}
