% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EvaluationFromExternalStats.R
\name{estimationSummaryToDF}
\alias{estimationSummaryToDF}
\title{Convert estimation summary to a data frame}
\usage{
estimationSummaryToDF(estimationSummary)
}
\arguments{
\item{estimationSummary}{a named list of estimation summary}
}
\value{
a data frame
}
\description{
Convert estimation summary to a data frame
}
