% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EvaluationFromExternalStats.R
\name{getEstimationFieldNames}
\alias{getEstimationFieldNames}
\title{Get estimation field names}
\usage{
getEstimationFieldNames()
}
\value{
a character vector of field names
}
\description{
Get estimation field names
}
