% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EvaluationFromExternalStats.R
\name{getPerformanceMeasures}
\alias{getPerformanceMeasures}
\title{Get performance measures}
\usage{
getPerformanceMeasures(y, p, w = NULL)
}
\arguments{
\item{y}{a binary outcome vector}

\item{p}{probabilities vector}

\item{w}{(optional) weight vector}
}
\value{
Performence measures....
}
\description{
get performance measure from a pair of binary outcome vector and model predictions. The observations
may be weighted by a weight vector
}
