% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preDiagnostics.R
\name{getPreDiagnosticsFieldNames}
\alias{getPreDiagnosticsFieldNames}
\title{Get field names of structured pre-diagnostic object}
\usage{
getPreDiagnosticsFieldNames()
}
\value{
a character vector of field names
}
\description{
Get field names of structured pre-diagnostic object
}
