% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dualSGDWeightOptimizer.R
\name{getTunedOptimizer}
\alias{getTunedOptimizer}
\title{Get a tuned optimizer}
\usage{
getTunedOptimizer(wOptimizer, Z, mu, balanceTrain)
}
\arguments{
\item{wOptimizer}{optimization object of type \code{sgdTunedWeightOptimizer}}

\item{Z}{feature matrix}

\item{mu}{statistics vector}

\item{balanceTrain}{balancing object}
}
\value{
tuned optimizer
}
\description{
select a step size that gives the best objective after a few
starting iterations
}
