% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/postDiagnostics.R
\name{getWeightingResultsFieldNames}
\alias{getWeightingResultsFieldNames}
\title{Get weighting results field names}
\usage{
getWeightingResultsFieldNames()
}
\value{
a character vector of field names
}
\description{
Get weighting results field names
}
