% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mseWeightOptimizer.R
\name{optimizeSEWeightsTuned}
\alias{optimizeSEWeightsTuned}
\title{Optimization using minimum squeread error criterion with initial tuning of the
base learning rate.}
\usage{
optimizeSEWeightsTuned(wOptimizer, Z, mu)
}
\arguments{
\item{wOptimizer}{optimizer object of type \code{seTunedWeightOptimizer}}

\item{Z}{feature matrix}

\item{mu}{external expectations vector}
}
\value{
optimization results. A list
}
\description{
Optimization using minimum squeread error criterion with initial tuning of the
base learning rate.
}
