% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EvaluationFromExternalStats.R
\name{postDiagnostics}
\alias{postDiagnostics}
\title{Post reweighing diagnostics}
\usage{
postDiagnostics(w, z, mu)
}
\arguments{
\item{w}{a vector of weights}

\item{z}{a data frame of transformed feature-outcome pairs}

\item{mu}{a vector of means of transformed feature-outcome pairs}
}
\value{
a named list with the following:
maxw
TODO   W2
kl
}
\description{
compute diagnostics of weighted samples
}
