% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preDiagnostics.R
\name{preCheckBinaryFeatures}
\alias{preCheckBinaryFeatures}
\title{pre check binary features}
\usage{
preCheckBinaryFeatures(z, mu, maxSubset = 20000)
}
\arguments{
\item{z}{data frame of features}

\item{mu}{named vectors of statistics}

\item{maxSubset}{max number of samples in training subsets}

\item{includeFeatures}{indicators}
}
\value{
a list
}
\description{
pre check binary features
}
