% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EvaluationFromExternalStats.R
\name{randomSubsetSampler}
\alias{randomSubsetSampler}
\title{Random subset sampler}
\usage{
randomSubsetSampler(n, nmax, replace = F)
}
\arguments{
\item{n}{number of samples}

\item{nmax}{maximum number of subsamples}

\item{replace}{boolean for sampling with replacement}
}
\value{
a \code{randomSubsetSampler} objects
}
\description{
Random subset sampler
}
