% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reweightTransforms.R
\name{reweightTransfrom}
\alias{reweightTransfrom}
\title{a re-weight transform that supports interactions among features}
\description{
a class that generates transform formulas
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{outcomeCol}}{name of outcome column}

\item{\code{interactionVars}}{a character or character vector with second order interaction variables. For every variable
in this field, an interaction is added with all other variables.}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-reweightTransfrom-new}{\code{reweightTransfrom$new()}}
\item \href{#method-reweightTransfrom-getFormula}{\code{reweightTransfrom$getFormula()}}
\item \href{#method-reweightTransfrom-clone}{\code{reweightTransfrom$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-reweightTransfrom-new"></a>}}
\if{latex}{\out{\hypertarget{method-reweightTransfrom-new}{}}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{reweightTransfrom$new(outcomeCol = "Y", interactionVars = NULL)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{outcomeCol}}{outcome column name}

\item{\code{interactionVars}}{names of interaction variables (NULL, character or character vector)}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-reweightTransfrom-getFormula"></a>}}
\if{latex}{\out{\hypertarget{method-reweightTransfrom-getFormula}{}}}
\subsection{Method \code{getFormula()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{reweightTransfrom$getFormula(X)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{X}}{a data frame in numeric format}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-reweightTransfrom-clone"></a>}}
\if{latex}{\out{\hypertarget{method-reweightTransfrom-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{reweightTransfrom$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
