% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mseWeightOptimizer.R
\name{seTunedWeightOptimizer}
\alias{seTunedWeightOptimizer}
\title{Optimization using minimum squared error criterion with initial tuning of the
base learning rate.}
\usage{
seTunedWeightOptimizer(
  alphas,
  minSd = 1e-04,
  nu0 = NULL,
  nTuneIter = 50,
  nIter = 1000,
  outputDir = NULL,
  improveTh = 1e-04,
  momentumC = 0.9,
  approxUpdate = F,
  maxErr = 0.01
)
}
\arguments{
\item{maxErr}{}
}
\value{
an object of class seTunedWeightOptimizer
}
\description{
Optimization using minimum squared error criterion with initial tuning of the
base learning rate.
}
