% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dualSGDWeightOptimizer.R
\name{sgdTunedWeightOptimizer}
\alias{sgdTunedWeightOptimizer}
\title{Optimization using minimum squared error criterion with initial tuning of the
base learning rate.}
\usage{
sgdTunedWeightOptimizer(
  epsilon0s,
  nTuneIter = 50,
  minSd = 1e-04,
  nu0 = NULL,
  nIter = 1000,
  batchSize = 20000,
  normalizeNu = T,
  momentumC = 0.9,
  outputDir = NULL,
  polyBeta = 1,
  maxAbsNu = 16,
  maxNuNorm = 10,
  improveTh = 1e-04,
  nProbe = 50,
  averaging = F
)
}
\arguments{
\item{nTuneIter}{number of iteration for finding the correct epsilon}

\item{minSd}{minimum standard deviation for removing features}

\item{nu0}{initial value}

\item{nIter}{maximum number of iterations}

\item{batchSize}{batch size}

\item{momentumC}{momentum coefficient}

\item{outputDir}{output directory}

\item{polyBeta}{coefficient for learning rate decay}

\item{maxAbsNu}{maximum absolute value of nu to avoid numeric overflow}

\item{maxNuNorm}{regulaziation parameter to limit max nu norm}

\item{improveTh}{threshold for stopping}

\item{nProbe}{number of iterations for showing intermediate resualts and assessments}

\item{averaging}{average the parameters over iterations}

\item{epsilon0}{learning rates vector}

\item{normalizeMu}{force Nu to represent a distribution}
}
\value{
an object of class seTunedWeightOptimizer
}
\description{
Optimization using minimum squared error criterion with initial tuning of the
base learning rate.
}
