% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dualSGDWeightOptimizer.R
\name{stochasticWeightOptimizer}
\alias{stochasticWeightOptimizer}
\title{'Base class' of sgdWeightOptimizer}
\usage{
stochasticWeightOptimizer(
  minSd = 1e-04,
  nu0 = NaN,
  nIter = 1000,
  epsilon0 = 0.1,
  batchSize = 10000
)
}
\description{
'Base class' of sgdWeightOptimizer
}
