% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EvaluationFromExternalStats.R
\name{stratifiedSampler}
\alias{stratifiedSampler}
\title{Stratified subset sampler}
\usage{
stratifiedSampler(n, nmax, Y)
}
\arguments{
\item{n}{number of samples}

\item{nmax}{maximum number of samples}

\item{Y}{outcome vector}
}
\value{
a \code{stratifiedSampler} objects
}
\description{
Stratified subset sampler
}
