% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transformClassifierData.R
\name{transformClassifierData}
\alias{transformClassifierData}
\title{Transform classifier data}
\usage{
transformClassifierData(
  d,
  transformType = "Table 1",
  interactionVar = NULL,
  outcomeCol = "Y",
  outcomeBalance = outcomeBalance
)
}
\arguments{
\item{d}{data frame}

\item{transformType}{\code{'Table 1'} or \code{'Interaction'} or \code{'Flat'}}

\item{outcomeCol}{string specifying the name of the outcome column}

\item{outcomeBalance}{boolean specifying if to use outcome in balancing}

\item{interactionVars}{interaction variables in case \code{transformType} is \code{'Interaction'}}
}
\value{
A data frame
}
\description{
Transform features and outcome columns, for example concatenate squared features, feature-outcome interaction
and more.
}
