% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EvaluationFromExternalStats.R
\name{warmStart}
\alias{warmStart}
\title{Initialize parameters using a short run on the entire set}
\usage{
warmStart(externalEstimatorSettings, internalData, externalStats)
}
\arguments{
\item{externalEstimatorSettings}{an object of class \code{externalEstimatorSettings} with the parameters
of the reweighing algorithm}

\item{internalData}{a list that includes internal data and predictions with the following fields:
z: a data frame of transformed feature-outcome pairs
y: a vector of outcomes
p: a vector of predicted outcome probabilities}

\item{externalStats}{a vector of means of transformed feature-outcome pairs}
}
\value{
an object of class \code{externalEstimatorSettings}
}
\description{
Initialize parameters using a short run on the entire set
}
