% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/WeightedEvaluationSummary.R
\name{WeightedLogLike}
\alias{WeightedLogLike}
\title{Compute weighted log-likelihood}
\usage{
WeightedLogLike(Y, p, w = NULL)
}
\arguments{
\item{Y}{a vector of binary labels}

\item{p}{a vector of numeric predictions (assuming in the range \link{0,1})}

\item{w}{a vector of non-negative weights}
}
\value{
weighted log-likelihood
}
\description{
compute weighted log-likelihood of a binary classifier predictions with respect to true labels
}
