% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EvaluationFromExternalStats.R
\name{createExternalEstimatorSettings}
\alias{createExternalEstimatorSettings}
\title{Create settings for \code{estimateExternalPerformanceFromStatistics}}
\usage{
createExternalEstimatorSettings(
  reweightAlgorithm,
  nMaxReweight = 20000,
  nRepetitions = 1,
  stratifiedSampling = T,
  maxDiff = 0.01,
  maxWSMD = 0.1,
  outputDir = getwd(),
  maxCores = 1,
  shortName = NULL,
  warmStartAlgorithm = NULL
)
}
\arguments{
\item{reweightAlgorithm}{algorithm that attempts to reweight an internal sample to get similar expectations from an
external one}

\item{nMaxReweight}{maximum number of samples for re-weighting. If this number is smaller than the number of
observations, then the estimator acts on sub-samples of the data-set.}

\item{nRepetitions}{number of repetitions with sub-samples.}

\item{stratifiedSampling}{attempt to sample a balanced number of samples for both outcome groups.}

\item{maxDiff}{maximum difference between  external and internal fequencies in case of unary variables or
binary variables with high proportion.}

\item{maxWSMD}{maximum allowed weighted standardized mean difference}

\item{outputDir}{output directory for logging}

\item{maxCores}{maximum number of cores for parallel processing of multiple sub-samples.}

\item{shortName}{a short name for display puposes}

\item{warmStartAlgorithm}{algorithm for intialization of weights before bootstrapping}
}
\value{
An object of class \code{externalEstimatorSettings}

TODO update this
}
\description{
evaluation from external stats
}
