% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EvaluationFromExternalStats.R
\name{estimateExternalPerformanceFromStatistics}
\alias{estimateExternalPerformanceFromStatistics}
\title{Estimate external model performance from statistics and internal
test data}
\usage{
estimateExternalPerformanceFromStatistics(
  internalData,
  externalStats,
  externalEstimatorSettings,
  createEstimationLogger = T
)
}
\arguments{
\item{internalData}{a list that includes internal data and predictions with the following fields:
z: a data frame of transformed feature-outcome pairs
y: a vector of outcomes
p: a vector of predicted outcome probabilities}

\item{externalStats}{a vector of means of transformed feature-outcome pairs

Reweighing algorithm parameters:}

\item{externalEstimatorSettings}{an object of class \code{externalEstimatorSettings}}

\item{createEstimationLogger}{create a logger in outputDirectory}
}
\value{
an object of \code{estimatedExternalPerformanceFromStatistics} with the following fields:

\itemize{
\item{\code{status}:} {execution status, either \code{'Success'} or \code{'Failure'}}
\item{\code{preDiagnosis}:} {Pre-diagnosis results object}
\item{\code{estimationTime}:} {execution time}
\item{\code{estimation}:} {a data-frame with a single column \code{'value'} and row-names that determine various
estimators.}
}
}
\description{
Estimate external performance using external statistics and an internal dataset.
This function reweights the internal \code{z} matrix with the objective to make the weighted means as close as
possible to the external means represented by \code{externalStats}. Performance measures are estimated using the
resulting weights, the vector of actual outcomes \code{y}, and the predicted outcome probabilities \code{p}.
}
