% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EvaluationFromExternalStats.R
\name{getHighlySkewedBinaryFeatures}
\alias{getHighlySkewedBinaryFeatures}
\title{Get highly skewed binary features}
\usage{
getHighlySkewedBinaryFeatures(mu, z, maxDiff)
}
\arguments{
\item{mu}{expectations}

\item{z}{data}

\item{maxDiff}{max difference from with to check max proportion}
}
\description{
Assuming that if the max proportion is violated than the expectations may be small and therefore checking also
maxDiff
}
