% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preDiagnostics.R
\name{getHighlySkewedBinaryFeatures}
\alias{getHighlySkewedBinaryFeatures}
\title{Get highly skewed binary features}
\usage{
getHighlySkewedBinaryFeatures(mu, z, minNumReport = 20, maxDiff = 0.01)
}
\arguments{
\item{mu}{expectations}

\item{z}{data}

\item{minNumReport}{minimum number of samples to report in the logger

TODO better treatment cases in which binary values may not be 0 or 1. Transform to this form if needed}
}
\description{
Get highly skewed binary features
}
