% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dualSGDWeightOptimizer.R
\name{getTunedOptimizer}
\alias{getTunedOptimizer}
\title{Get a tuned optimizer}
\usage{
getTunedOptimizer(wOptimizer, cTrain, extendedMu, balanceTrain)
}
\arguments{
\item{balanceTrain}{}
}
\value{
tuned optimizer
}
\description{
select a step size that gives the best objective after a few
starting iterations
}
