% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LearningWithExternalStats.R
\docType{package}
\name{LearningWithExternalStats}
\alias{LearningWithExternalStats}
\title{LearningWithExternalStats}
\description{
A package for augmenting learning by utilizing external summary statistics in addition to internal
detailed data
}
