% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mseWeightOptimizer.R
\name{optimizeSEWeightsTuned}
\alias{optimizeSEWeightsTuned}
\title{Optimization using minimum squered error criterion with initial tuning of the
base learning rate.}
\usage{
optimizeSEWeightsTuned(wOptimizer, Z, mu)
}
\arguments{
\item{mu}{}
}
\value{
optimization results. A list
}
\description{
Optimization using minimum squered error criterion with initial tuning of the
base learning rate.
}
