% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dualSGDWeightOptimizer.R
\name{optimizeWeightSGD}
\alias{optimizeWeightSGD}
\title{Optimization using stochastic gradient descent}
\usage{
optimizeWeightSGD(
  wOptimizer,
  C,
  extendedMu,
  balanceTrain,
  cValidation = NULL,
  yValidation = NULL,
  verbose = F,
  improveTh = 1e-04,
  nProbe = 50,
  averaging = F
)
}
\arguments{
\item{verbose}{}
}
\value{
optimization results. A list
}
\description{
Optimization using stochastic gradient descent
}
