% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/postDiagnostics.R
\name{postDiagnostics}
\alias{postDiagnostics}
\title{Post weighting diagnostics}
\usage{
postDiagnostics(w, z, mu)
}
\arguments{
\item{w}{a vector of weights}

\item{z}{a data frame of transformed feature-outcome pairs}

\item{mu}{a vector of means of transformed feature-outcome pairs}
}
\value{
a named list with the following:
\itemize{
\item {\code{Max weight}:} {Maximal weight}
\item {\code{chi2ToUnif}:} {Chi squared divergence between weights and a uniform distribution}
\item {\code{kl}:} {KL distance to uniform distribution}
\item {\code{Max Weighted SMD}:} {Maximum over features of SMD between internal set and external statistics}
}
}
\description{
compute diagnostics of weighted samples
}
