% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EvaluationFromExternalStats.R
\name{preDiagnostics}
\alias{preDiagnostics}
\title{Pre re-weighting diagnostics}
\usage{
preDiagnostics(z, mu, maxDiff, npMinRatio = 4)
}
\arguments{
\item{z}{a data frame of transformed feature-outcome pairs}

\item{mu}{a vector of means of transformed feature-outcome pairs}

\item{maxDiff}{maximum difference from which to check max prop}

\item{npMinRatio}{minimum ratio between number of features and number of examples}
}
\value{
a named list with the following fields:
...
}
\description{
compare external expectations and internal means before reweighting
}
