% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dualSGDWeightOptimizer.R
\name{sgdTunedWeightOptimizer}
\alias{sgdTunedWeightOptimizer}
\title{Optimization using minimum squared error criterion with initial tuning of the
base learning rate.}
\usage{
sgdTunedWeightOptimizer(
  epsilon0s,
  nTuneIter = 50,
  minSd = 1e-04,
  nu0 = NULL,
  nIter = 1000,
  batchSize = 20000,
  normalizeNu = T,
  momentumC = 0.9,
  outputDir = NULL,
  polyBeta = 1,
  maxAbsNu = 16,
  maxNuNorm = 10,
  improveTh = 1e-04,
  nProbe = 50,
  averaging = F
)
}
\arguments{
\item{averaging}{}
}
\value{
an object of class seTunedWeightOptimizer
}
\description{
Optimization using minimum squared error criterion with initial tuning of the
base learning rate.
}
