% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EvaluationFromExternalStats.R
\name{summarizeBootstrap}
\alias{summarizeBootstrap}
\title{Summarize bootstrap}
\usage{
summarizeBootstrap(b)
}
\arguments{
\item{b}{bootstrap results matrix columns correspond to different metrix, rows to repetitions. Columns
should be named by the metric.}
}
\value{
a named list with bootstrap statistics for every metric
}
\description{
Summarize statistics of metrics obtained by bootstrapping
}
